+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/autoinstall/AbstractInstallStrategy.kt
@@ -15,7 +15,9 @@ abstract class AbstractInstallStrategy : ComponentMetadataRule {
 
     protected abstract val shouldInstallModule: Boolean
 
-    protected open val minSupportedThirdPartyVersion: SemVer = SemVer(0, 0, 0)
+    protected open val minSupportedThirdPartyVersion: SemVer? = null
+
+    protected open val maxSupportedThirdPartyVersion: SemVer? = null
 
     protected open val minSupportedSentryVersion: SemVer = SemVer(0, 0, 0)
 
@@ -27,13 +29,27 @@ abstract class AbstractInstallStrategy : ComponentMetadataRule {
             }
             return
         }
-        val thirdPartySemVersion = SemVer.parse(context.details.id.version)
-        if (thirdPartySemVersion < minSupportedThirdPartyVersion) {
-            logger.warn {
-                "$sentryModuleId won't be installed because the current version is " +
-                    "lower than the minimum supported version ($minSupportedThirdPartyVersion)"
-            }
-            return
+        minSupportedThirdPartyVersion?.let {
+            parseVersion(context.details.id.version)?.let { thirdPartySemVersion ->
+                if (thirdPartySemVersion < it) {
+                    logger.warn {
+                        "$sentryModuleId won't be installed because the current version is " +
+                            "lower than the minimum supported version ($it)"
+                    }
+                    return
+                }
+            } ?: return
+        }
+        maxSupportedThirdPartyVersion?.let {
+            parseVersion(context.details.id.version)?.let { thirdPartySemVersion ->
+                if (thirdPartySemVersion > it) {
+                    logger.warn {
+                        "$sentryModuleId won't be installed because the current version is " +
+                            "higher than the maximum supported version ($it)"
+                    }
+                    return
+                }
+            } ?: return
         }
 
         if (minSupportedSentryVersion.major > 0) {
@@ -66,6 +82,20 @@ abstract class AbstractInstallStrategy : ComponentMetadataRule {
                     "$sentryModuleId was successfully installed with version: $sentryVersion"
                 }
             }
+        }
+    }
+
+    private fun parseVersion(version: String): SemVer? {
+        // older Spring versions ended in .RELEASE
+        return parseVersionSafely(version.removeSuffix(".RELEASE"))
+    }
+
+    private fun parseVersionSafely(version: String): SemVer? {
+        try {
+            return SemVer.parse(version)
+        } catch (t: Throwable) {
+            logger.warn { "Unable to parse version $version as a semantic version." }
+            return null
         }
     }
 }